requirements.txt
Django==4.2.25
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-decouple==3.8
Pillow==11.3.0
django-filter==24.2
